uint8 bilinear on LLVM with atol=1

+++ tinygrad/tensor.py
@@ -2038,7 +2038,7 @@ class Tensor:
     assert isinstance(size, (tuple,list)) and all_int(size) and 0 < len(size) <= self.ndim, f"invalid {size=}"
     assert mode in ["linear", "nearest", "nearest-exact"], "only supports linear, nearest, and nearest-exact interpolate"
     if mode != "linear" and align_corners: raise ValueError("align_corners can only be set with interpolate mode linear")
-    x, expand = self, list(self.shape)
+    x, expand = self.cast(dtypes.float32), list(self.shape)
     for i in range(-len(size), 0):
       scale = (self.shape[i] - int(align_corners)) / (size[i] - int(align_corners))
       arr, reshape = Tensor.arange(size[i], dtype=dtypes.float32, device=self.device), [1] * self.ndim
